Refactor Axis and Side into common packages
